django/comarques/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
html5lib==1.1
idna==3.10
inflection==0.5.1
lxml==5.4.0
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.2.1
psycopg==3.2.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.4.0
webencodings==0.5.1
